+++ wallet/src/main/scala/org/bitcoins/wallet/api/UnlockWalletResult.scala
@@ -10,4 +10,9 @@ sealed trait UnlockWalletError extends Error with UnlockWalletResult
 object UnlockWalletError {
   final case object BadPassword
       extends Error("Bad password for unlocking wallet!")
+      with UnlockWalletError
+
+  final case class JsonParsingError(message: String)
+      extends Error(message)
+      with UnlockWalletError
 }
