+++ wallet/src/main/scala/org/bitcoins/wallet/db/WalletDbManagement.scala
@@ -4,7 +4,6 @@ import org.bitcoins.db.DbManagement
 import org.bitcoins.wallet.models.{
   AccountTable,
   AddressTable,
-  MnemonicCodeTable,
   UTXOSpendingInfoTable
 }
 import slick.jdbc.SQLiteProfile.api._
@@ -12,11 +11,10 @@ import slick.jdbc.SQLiteProfile.api._
 sealed abstract class WalletDbManagement extends DbManagement {
   private val accountTable = TableQuery[AccountTable]
   private val addressTable = TableQuery[AddressTable]
-  private val mnemonicDAO = TableQuery[MnemonicCodeTable]
   private val utxoDAO = TableQuery[UTXOSpendingInfoTable]
 
   override val allTables: List[TableQuery[_ <: Table[_]]] =
-    List(accountTable, addressTable, mnemonicDAO, utxoDAO)
+    List(accountTable, addressTable, utxoDAO)
 
 }
